Split callback implementations into relevant files
