+++ .coveragerc
@@ -2,3 +2,4 @@
 omit =
     datadoc/tests/*
     datadoc/__init__.py
+    datadoc/frontend/callbacks/register.py
