+++ datadoc/tests/test_callbacks.py
@@ -6,7 +6,18 @@ from datadoc_model.Model import DataDocDataSet, DataDocVariable, LanguageStrings
 
 import datadoc.state as state
 from datadoc.backend.DataDocMetadata import DataDocMetadata
-from datadoc.frontend.callbacks import register
+from datadoc.frontend.callbacks.dataset import (
+    accept_dataset_metadata_input,
+    change_language_dataset_metadata,
+    update_dataset_metadata_language,
+    update_global_language_state,
+)
+from datadoc.frontend.callbacks.utils import find_existing_language_string
+from datadoc.frontend.callbacks.variables import (
+    accept_variable_metadata_input,
+    update_variable_table_dropdown_options_for_language,
+    update_variable_table_language,
+)
 from datadoc.frontend.fields.DisplayDataset import DISPLAYED_DROPDOWN_DATASET_ENUMS
 from datadoc.frontend.fields.DisplayVariables import VariableIdentifiers
 from datadoc.tests.utils import TEST_PARQUET_FILEPATH
@@ -42,7 +53,7 @@ LANGUAGE_OBJECT = LanguageStrings(en=ENGLISH_NAME, nb=BOKMÅL_NAME)
 
 def test_accept_variable_metadata_input_no_change_in_data():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
-    output = register.accept_variable_metadata_input(DATA_ORIGINAL, DATA_ORIGINAL)
+    output = accept_variable_metadata_input(DATA_ORIGINAL, DATA_ORIGINAL)
     assert output[0] == DATA_ORIGINAL
     assert output[1] is False
     assert output[2] == ""
@@ -50,7 +61,7 @@ def test_accept_variable_metadata_input_no_change_in_data():
 
 def test_accept_variable_metadata_input_new_data():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
-    output = register.accept_variable_metadata_input(DATA_VALID, DATA_ORIGINAL)
+    output = accept_variable_metadata_input(DATA_VALID, DATA_ORIGINAL)
 
     assert state.metadata.variables_lookup["pers_id"].variable_role == "IDENTIFIER"
     assert output[0] == DATA_VALID
@@ -61,7 +72,7 @@ def test_accept_variable_metadata_input_new_data():
 def test_accept_variable_metadata_input_incorrect_data_type():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
     previous_metadata = deepcopy(state.metadata.meta.variables)
-    output = register.accept_variable_metadata_input(DATA_INVALID, DATA_ORIGINAL)
+    output = accept_variable_metadata_input(DATA_INVALID, DATA_ORIGINAL)
 
     assert output[0] == DATA_ORIGINAL
     assert output[1] is True
@@ -71,9 +82,7 @@ def test_accept_variable_metadata_input_incorrect_data_type():
 
 def test_accept_dataset_metadata_input_new_data():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
-    output = register.accept_dataset_metadata_input(
-        DatasetState.INPUT_DATA, "dataset_state"
-    )
+    output = accept_dataset_metadata_input(DatasetState.INPUT_DATA, "dataset_state")
     assert output[0] is False
     assert output[1] == ""
     assert state.metadata.meta.dataset.dataset_state == "INPUT_DATA"
@@ -81,7 +90,7 @@ def test_accept_dataset_metadata_input_new_data():
 
 def test_accept_dataset_metadata_input_incorrect_data_type():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
-    output = register.accept_dataset_metadata_input(3.1415, "dataset_state")
+    output = accept_dataset_metadata_input(3.1415, "dataset_state")
     assert output[0] is True
     assert "validation error for DataDocDataSet" in output[1]
 
@@ -90,11 +99,11 @@ def test_update_dataset_metadata_language_strings():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
     state.metadata.meta.dataset.name = LANGUAGE_OBJECT
     state.current_metadata_language = SupportedLanguages.NORSK_BOKMÅL
-    output = register.update_dataset_metadata_language()
+    output = update_dataset_metadata_language()
     assert ENGLISH_NAME not in output
     assert BOKMÅL_NAME in output
     state.current_metadata_language = SupportedLanguages.ENGLISH
-    output = register.update_dataset_metadata_language()
+    output = update_dataset_metadata_language()
     assert ENGLISH_NAME in output
     assert BOKMÅL_NAME not in output
 
@@ -103,12 +112,12 @@ def test_update_dataset_metadata_language_enums():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
     state.metadata.meta.dataset.dataset_state = DatasetState.PROCESSED_DATA
     state.current_metadata_language = SupportedLanguages.NORSK_BOKMÅL
-    output = register.update_dataset_metadata_language()
+    output = update_dataset_metadata_language()
     assert DatasetState.PROCESSED_DATA.language_strings.en not in output
     assert DatasetState.PROCESSED_DATA.language_strings.nb not in output
     assert DatasetState.PROCESSED_DATA.name in output
     state.current_metadata_language = SupportedLanguages.ENGLISH
-    output = register.update_dataset_metadata_language()
+    output = update_dataset_metadata_language()
     assert DatasetState.PROCESSED_DATA.language_strings.en not in output
     assert DatasetState.PROCESSED_DATA.language_strings.nb not in output
     assert DatasetState.PROCESSED_DATA.name in output
@@ -117,7 +126,7 @@ def test_update_dataset_metadata_language_enums():
 def test_find_existing_language_string_no_existing_strings():
     dataset_metadata = DataDocDataSet()
     state.current_metadata_language = SupportedLanguages.NORSK_BOKMÅL
-    language_strings = register.find_existing_language_string(
+    language_strings = find_existing_language_string(
         dataset_metadata, BOKMÅL_NAME, "name"
     )
     assert language_strings == LanguageStrings(nb=BOKMÅL_NAME)
@@ -127,7 +136,7 @@ def test_find_existing_language_string_pre_existing_strings():
     dataset_metadata = DataDocDataSet()
     dataset_metadata.name = LANGUAGE_OBJECT
     state.current_metadata_language = SupportedLanguages.NORSK_NYNORSK
-    language_strings = register.find_existing_language_string(
+    language_strings = find_existing_language_string(
         dataset_metadata, NYNORSK_NAME, "name"
     )
     assert language_strings == LanguageStrings(
@@ -139,7 +148,7 @@ def test_update_variable_table_language():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
     test_variable = random.choice([v.short_name for v in state.metadata.meta.variables])
     state.metadata.variables_lookup[test_variable].name = LANGUAGE_OBJECT
-    output = register.update_variable_table_language(
+    output = update_variable_table_language(
         [v.dict() for v in state.metadata.meta.variables],
         SupportedLanguages.NORSK_BOKMÅL,
     )
@@ -155,13 +164,13 @@ def test_nonetype_value_for_language_string():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
     state.metadata.variables_lookup["pers_id"].name = LANGUAGE_OBJECT
     state.current_metadata_language = SupportedLanguages.NORSK_NYNORSK
-    register.accept_variable_metadata_input(DATA_NONETYPE, DATA_ORIGINAL)
+    accept_variable_metadata_input(DATA_NONETYPE, DATA_ORIGINAL)
 
     assert state.metadata.variables_lookup["pers_id"].name == LANGUAGE_OBJECT
 
 
 def test_update_variable_table_dropdown_options_for_language():
-    options = register.update_variable_table_dropdown_options_for_language(
+    options = update_variable_table_dropdown_options_for_language(
         SupportedLanguages.NORSK_BOKMÅL
     )
     assert all(k in DataDocVariable.__fields__.keys() for k in options.keys())
@@ -178,13 +187,13 @@ def test_update_variable_table_dropdown_options_for_language():
 
 def test_update_global_language_state():
     language: SupportedLanguages = random.choice(list(SupportedLanguages))
-    register.update_global_language_state(language)
+    update_global_language_state(language)
     assert state.current_metadata_language == language
 
 
 def test_change_language_dataset_metadata():
     state.metadata = DataDocMetadata(TEST_PARQUET_FILEPATH)
-    value = register.change_language_dataset_metadata(SupportedLanguages.NORSK_NYNORSK)
+    value = change_language_dataset_metadata(SupportedLanguages.NORSK_NYNORSK)
     test = random.choice(DISPLAYED_DROPDOWN_DATASET_ENUMS)
     assert isinstance(value, tuple)
